Fix is_same_comparison for objects with list of dicts
Check mode would in most cases register changes when an object has a list of dicts longer than one item, for example the interfaces in a firewall zone. I've added a counter to allow the logic to compare the right items. It will probably fail when the order of the items is different.