Fix find_current_values diff logic for lists of dicts
Properly comparing the current and "filtered" objects is difficult when we encounter lists of dicts, as we have no way of telling which dict in the list whould be compared to any other dict in the other list. A primary key would help but will be different for each type of object. Is this what the q_origin_key was intended to do? I can't find any documentation on this